backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.3.2
duckduckgo-search==6.1.0
beautifulsoup4==4.12.2
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.0
httpx==0.25.2
python-dotenv==1.0.0
lxml==4.9.3
